requirements.txt
djangorestframework
pytest-asyncio
django
httpx
pytest
aiofiles
aiocsv
